+++ projects/ui-kit/src/lib/ui-kit/dynamic-selectors/components/dynamic-input/rtui-dynamic-input.component.ts
@@ -37,7 +37,7 @@ import { MatInput, MatLabel } from '@angular/material/input';
 import { noop } from 'rxjs';
 import { distinctUntilChanged, filter } from 'rxjs/operators';
 
-import { BlockDirective, ElemDirective } from '@rt-tools/core';
+import { BlockDirective, ConcatClassesPipe, ElemDirective } from '@rt-tools/core';
 import { INullable } from '@rt-tools/utils';
 import { areArraysEqual, transformArrayInput, transformStringInput } from '@rt-tools/utils';
 import {
@@ -60,12 +60,16 @@ interface FormModel {
 })
 export class RtuiDynamicInputAdditionalControlDirective {}
 
+const BEM_BLOCK: string = 'rtui-dynamic-input';
+
 @Component({
     selector: 'rtui-dynamic-input',
+    host: { class: BEM_BLOCK },
     templateUrl: './rtui-dynamic-input.component.html',
     styleUrls: ['./rtui-dynamic-input.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
     imports: [
+        ConcatClassesPipe,
         NgTemplateOutlet,
         ReactiveFormsModule,