+++ projects/ui-kit/src/lib/ui-kit/table/components/table-header-filter-cell/table-header-filter-cell.component.ts
@@ -28,7 +28,7 @@ import {
     transformArrayInput,
     isString,
 } from '@rt-tools/utils';
-import { RtIconOutlinedDirective } from '@rt-tools/core';
+import { ConcatClassesPipe, RtIconOutlinedDirective } from '@rt-tools/core';
 import { ITable, TABLE_COLUMN_FILTER_TYPES_ENUM } from '../../util/table-column.interface';
 import { MatFormField, MatFormFieldAppearance, MatSuffix } from '@angular/material/form-field';
 import { MatInput } from '@angular/material/input';
@@ -42,13 +42,17 @@ import { RtuiClearButtonComponent } from '../clear-search-button/rtui-clear-butt
 import { BlockDirective } from '@rt-tools/core';
 import { MatTooltip } from '@angular/material/tooltip';
 
+const BEM_BLOCK: string = 'rtui-table-header-filter-cell';
+
 @Component({
     selector: 'rtui-table-header-filter-cell',
+    host: { class: BEM_BLOCK },
     templateUrl: './table-header-filter-cell.component.html',
     styleUrls: ['./table-header-filter-cell.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
     providers: [provideNativeDateAdapter()],
     imports: [
+        ConcatClassesPipe,
         FormsModule,
 
         // Material